[Refactor](BE) Uniform Expr, Block, Columns size-relative interface (Part I) (#42930)
## Proposed changes
In this pr we refactored some base interface to avoid narrow cast with
data consistent kept.

We uniformed these type of numbers:
```
columns in Block        - U32 (just use no more than U16)
Rows in Block           - U64 (others U32, ColumnString64 U64)
column_id in expr arg   - I32 (because -1 is meaningful)
child of Expr           - U16 (FE use int32, have a restriction of 10000. adjustable but no guarantee for result)
```

So, for functions' argument column id, we use `uint32` to replace
`size_t`.
For rows, we use `size_t` to replace some `int`.
For child of expr, we cast it as `uint_16` to use, rather than `size_t`
as `vector::size()`

Refactor of table functions: https://github.com/apache/doris/pull/43218

+++ be/src/exprs/bitmapfilter_predicate.h
@@ -37,7 +37,7 @@ public:
     virtual void light_copy(BitmapFilterFuncBase* other) { _not_in = other->_not_in; }
     virtual uint16_t find_fixed_len_olap_engine(const char* data, const uint8* nullmap,
                                                 uint16_t* offsets, int number) = 0;
-    virtual void find_batch(const char* data, const uint8* nullmap, int number,
+    virtual void find_batch(const char* data, const uint8* nullmap, size_t number,
                             uint8* results) const = 0;
     virtual size_t size() const = 0;
     bool is_not_in() const { return _not_in; }
@@ -65,7 +65,7 @@ public:
     uint16_t find_fixed_len_olap_engine(const char* data, const uint8* nullmap, uint16_t* offsets,
                                         int number) override;
 
-    void find_batch(const char* data, const uint8* nullmap, int number,
+    void find_batch(const char* data, const uint8* nullmap, size_t number,
                     uint8* results) const override;
 
     bool empty() override { return _bitmap_value->empty(); }
@@ -133,9 +133,9 @@ uint16_t BitmapFilterFunc<type>::find_fixed_len_olap_engine(const char* data, co
 }
 
 template <PrimitiveType type>
-void BitmapFilterFunc<type>::find_batch(const char* data, const uint8* nullmap, int number,
+void BitmapFilterFunc<type>::find_batch(const char* data, const uint8* nullmap, size_t number,
                                         uint8* results) const {
-    for (int i = 0; i < number; i++) {
+    for (size_t i = 0; i < number; i++) {
         results[i] = false;
         if (nullmap != nullptr && nullmap[i]) {
             continue;
